feat: authorization

+++ src/users/users.controller.ts
@@ -1,35 +1,77 @@
-import { Body, Controller, Delete, Get, Param, Post, Query, } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Headers, HttpCode, HttpException, HttpStatus, Param, Post, Query, Req, UseGuards, } from '@nestjs/common';
 import { UsersService } from './users.service';
-import { UserCreateDto, UserUpdateDto } from '../dto/users.dto';
+import { UserCreateDto, UserLoginDto, UserUpdateDto } from '../dto/users.dto';
+import { AuthGuard } from '@nestjs/passport';
+import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import { checkPermission } from 'src/ultiService/ultiService';
+import { failCode, unAuthor } from 'src/config/respone.service';
 
+@ApiTags("UsersService")
 @Controller('users')
 export class UsersController {
     constructor(
         private readonly UsersService: UsersService
     ) { }
 
-    @Get('get')
-    getUser(@Query("keySearch") keySearch: string, @Query("pageNumber") pageNumber: string, @Query("pageSize") pageSize: string): Promise<any> {
-        return this.UsersService.getUsers(keySearch, parseInt(pageNumber), parseInt(pageSize))
-    };
-
     @Post('signup')
-    SignUp(@Body() UserInfo: UserCreateDto) {
+    signUp(@Body() UserInfo: UserCreateDto) {
         return this.UsersService.signup(UserInfo)
     }
 
+    @Post('login')
+    login(@Body() UserLogin: UserLoginDto) {
+        return this.UsersService.login(UserLogin)
+    }
+
+    @ApiBearerAuth()
+    @UseGuards(AuthGuard("jwt"))
+    @HttpCode(201)
+    @Get('get')
+    getUser(@Query("keySearch") keySearch: string, @Query("pageNumber") pageNumber: string, @Query("pageSize") pageSize: string, @Req() req) {
+        try {
+            if (!checkPermission(req, ["admin"])) return unAuthor()
+            return this.UsersService.getUsers(keySearch, parseInt(pageNumber), parseInt(pageSize))
+        } catch (error) {
+            throw new HttpException("Error Server", HttpStatus.INTERNAL_SERVER_ERROR)
+        }
+    };
+
+    @ApiBearerAuth()
+    @UseGuards(AuthGuard("jwt"))
+    @HttpCode(201)
     @Post("update/:id")
-    Update(@Body() UserInfo: UserUpdateDto, @Param("id") id: string) {
-        return this.UsersService.update(UserInfo, parseInt(id))
+    update(@Body() UserInfo: UserUpdateDto, @Param("id") id: string, @Req() req) {
+        try {
+            if (!checkPermission(req, ["admin", "toureguide"])) return unAuthor()
+            return this.UsersService.update(UserInfo, parseInt(id))
+        } catch (error) {
+            throw new HttpException("Error Server", HttpStatus.INTERNAL_SERVER_ERROR)
+        }
     }
 
+    @ApiBearerAuth()
+    @UseGuards(AuthGuard("jwt"))
+    @HttpCode(201)
     @Get('get-id')
-    GetOne(@Query("id") id: string) {
-        return this.UsersService.getOne(parseInt(id))
+    getOne(@Query("id") id: string, @Req() req) {
+        try {
+            if (!checkPermission(req, ["admin", "toureguide"])) return unAuthor()
+            return this.UsersService.getOne(parseInt(id))
+        } catch (error) {
+            throw new HttpException("Error Server", HttpStatus.INTERNAL_SERVER_ERROR)
+        }
     }
 
+    @ApiBearerAuth()
+    @UseGuards(AuthGuard("jwt"))
+    @HttpCode(201)
     @Delete('delete')
-    Delete(@Query("id") id: string) {
-        return this.UsersService.delete(parseInt(id))
+    delete(@Query("id") id: string, @Req() req) {
+        try {
+            if (!checkPermission(req, ["admin", "toureguide"])) return unAuthor()
+            return this.UsersService.delete(parseInt(id))
+        } catch (error) {
+            throw new HttpException("Error Server", HttpStatus.INTERNAL_SERVER_ERROR)
+        }
     }
 }
